chore(ui): Edit secrets file structure
Reorganize secret related files in anticipation of adding deeper
routes to make the structure clearer.
